backend/pyproject.toml
[project]
name = "engineering-signal-aggregator"
version = "0.1.0"
description = "AI-assisted engineering signal aggregator"
requires-python = ">=3.11"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy>=2.0",
  "psycopg[binary]",
  "pydantic",
  "python-dotenv",
  "alembic>=1.18.0",
  "httpx>=0.28.1",
]
